refactor: change location of cypress test codes
